+++ mido/backends/rtmidi.py
@@ -5,9 +5,18 @@ http://pypi.python.org/pypi/python-rtmidi/
 """
 from __future__ import absolute_import
 import os
+import sys
 import time
+
+PY2 = (sys.version_info.major == 2)
+if PY2:
+    import Queue as queue
+else:
+    import queue
+
 import rtmidi
-from ..ports import BaseInput, BaseOutput
+from ..ports import get_sleep_time, BaseInput, BaseOutput
+
 
 def _get_api_lookup():
     api_to_name = {}
@@ -62,10 +71,11 @@ class PortCommon(object):
 
         rtapi = _get_api_id(api)
         opening_input = hasattr(self, 'receive')
-        
+
         if opening_input:
             self._rt = rtmidi.MidiIn(rtapi=rtapi)
             self._rt.ignore_types(False, False, True)
+            self._queue = queue.Queue()
             self.callback = kwargs.get('callback')
         else:
             self._rt = rtmidi.MidiOut(rtapi=rtapi)
@@ -101,40 +111,85 @@ class PortCommon(object):
         if virtual:
             self._device_type = 'virtual {}'.format(self._device_type)
 
+    def _close(self):
+        self._rt.close_port()
+        del self._rt  # Virtual ports are closed when this is deleted.
+
+class Input(PortCommon, BaseInput):
     @property
     def callback(self):
         return self._callback
 
     @callback.setter
     def callback(self, func):
-        self._callback = func
-        if func is None:
-            self._rt.cancel_callback()
-        else:
-            self._rt.set_callback(self._callback_wrapper)
+        self._rt.cancel_callback()
 
-    def _callback_wrapper(self, message_data, data):
+        if func:
+            # First send all queued messages to the callback.
+            while True:
+                try:
+                    func(self._queue.get_nowait())
+                except queue.Empty:
+                    break
+
+        if func is None:
+            self._rt.set_callback(self._feed_queue)
+            self._callback = None
+        else:
+            self._rt.set_callback(self._feed_callback)
+            self._callback = func
+
+    def _feed_queue(self, message_data, data):
         self._parser.feed(message_data[0])
         for message in self._parser:
-            if self._callback:
-                self._callback(message)
+            self._queue.put(message)
 
-    def _close(self):
-        self._rt.close_port()
-        del self._rt  # Virtual ports are closed when this is deleted.
+    def _feed_callback(self, message_data, data):
+        self._parser.feed(message_data[0])
+        for message in self._parser:
+            self._callback(message)
 
-class Input(PortCommon, BaseInput):
-    def _receive(self, block=True):
-        # Since there is no blocking read in RtMidi, the block
-        # flag is ignored and the enclosing receive() takes care
-        # of blocking.
+    # receive() implementation:
+    #
+    # We need to override receive() here because _receive() doesn't allow
+    # us to return a messages directly and putting it in self._messages is
+    # not an option.
+    #
+    # Todo: add doc strings.
 
-        while True:
-            message_data = self._rt.get_message()
-            if message_data is None:
-                break
-            else:
-                self._parser.feed(message_data[0])
+    if PY2:
+        # In Python 2 queue.get() doesn't respond to CTRL-C. A workaroud is
+        # to call queue.get(timeout=100) (very high timeout) in a loop, but all
+        # that does is poll with a timeout of 50 milliseconds. This results in
+        # much too high latency.
+        #
+        # It's better to do our own polling with a shorter sleep time.
+        #
+        # See Issue #49 and https://bugs.python.org/issue8844
+
+        def receive(self, block=True):
+            sleep_time = get_sleep_time()
+            while True:
+                try:
+                    return self._queue.get_nowait()
+                except queue.Empty:
+                    if block:
+                        time.sleep(sleep_time)
+                        continue
+                    else:
+                        return None
+    else:
+
+        def receive(self, block=True):
+            try:
+                return self._queue.get(block=block)
+            except queue.Empty:
+                return None
+
+    receive.__doc__ = BaseInput.receive.__doc__
+
+    def pending(self):
+        return self._queue.qsize()
 
 class Output(PortCommon, BaseOutput):
     def _send(self, message):